refactor functions into ansible_runner.utils
This change moves functions from interface.py to utils.py including all
necessary import statements.  No additional functional changes have been
made to the code.
