+++ ansible_runner/utils.py
@@ -3,6 +3,132 @@ import base64
 import json
 import sys
 import re
+import os
+import fcntl
+import tempfile
+import hashlib
+
+from collections import Iterable, Mapping
+
+from six import string_types
+
+
+
+def isplaybook(obj):
+    '''
+    Inspects the object and returns if it is a playbook
+
+    Args:
+        obj (object): The object to be inspected by this function
+
+    Returns:
+        boolean: True if the object is a list and False if it is not
+    '''
+    return isinstance(obj, Iterable) and not isinstance(obj, string_types)
+
+
+def isinventory(obj):
+    '''
+    Inspects the object and returns if it is an inventory
+
+    Args:
+        obj (object): The object to be inspected by this function
+
+    Returns:
+        boolean: True if the object is an inventory dict and False if it is not
+    '''
+    if isinstance(obj, Mapping):
+        return True
+    elif isinstance(obj, string_types):
+        return os.path.exists(obj)
+    raise ValueError('invalid object type for inventory')
+
+
+def dump_artifact(obj, path, filename=None):
+    '''
+    Write the artifact to disk at the specified path
+
+    Args:
+        obj (string): The string object to be dumped to disk in the specified
+            path.  The artifact filename will be automatically created
+
+        path (string): The full path to the artifacts data directory.
+
+        filename (string, optional): The name of file to write the artifact to.
+            If the filename is not provided, then one will be generated.
+
+    Returns:
+        string: The full path filename for the artifact that was generated
+    '''
+    p_sha1 = None
+
+    if not os.path.exists(path):
+        os.makedirs(path)
+    else:
+        p_sha1 = hashlib.sha1()
+        p_sha1.update(obj)
+
+    if filename is None:
+        fd, fn = tempfile.mkstemp(dir=path)
+    else:
+        fn = os.path.join(path, filename)
+
+    if os.path.exists(fn):
+        c_sha1 = hashlib.sha1()
+        c_sha1.update(open(fn).read())
+
+    if not os.path.exists(fn) or p_sha1.hexdigest() != c_sha1.hexdigest():
+        lock_fp = os.path.join(path, '.artifact_write_lock')
+        lock_fd = os.open(lock_fp, os.O_RDWR | os.O_CREAT, 0o600)
+        fcntl.lockf(lock_fd, fcntl.LOCK_EX)
+
+        try:
+            with open(fn, 'w') as f:
+                f.write(str(obj))
+        finally:
+            fcntl.lockf(lock_fd, fcntl.LOCK_UN)
+
+    return fn
+
+
+def to_artifacts(kwargs):
+    '''
+    Introspect the kwargs and dump objects to disk
+    '''
+    try:
+        private_data_dir = kwargs.get('private_data_dir')
+        if not private_data_dir:
+            private_data_dir = tempfile.mkdtemp()
+            kwargs['private_data_dir'] = private_data_dir
+
+        for key in ('playbook', 'inventory'):
+            obj = kwargs.get(key)
+            if obj:
+                if key == 'playbook' and isplaybook(obj):
+                    path = os.path.join(private_data_dir, 'project')
+                    kwargs['playbook'] = dump_artifact(json.dumps(obj), path, 'main.json')
+
+                elif key == 'inventory' and isinventory(obj):
+                    path = os.path.join(private_data_dir, 'inventory')
+                    if isinstance(obj, Mapping):
+                        kwargs['inventory'] = dump_artifact(json.dumps(obj), path, 'hosts.json')
+                    else:
+                        kwargs['inventory'] = dump_artifact(obj, path, 'hosts')
+
+        for key in ('envvars', 'extravars', 'passwords', 'settings'):
+            obj = kwargs.get(key)
+            if obj:
+                path = os.path.join(private_data_dir, 'env')
+                dump_artifact(json.dumps(obj), path, filename=key)
+                kwargs.pop(key)
+
+        if 'ssh_key' in kwargs:
+            path = os.path.join(private_data_dir, 'ssh_key')
+            dump_artifact(str(obj), path, filename='ssh_key')
+            kwargs.pop('ssh_key')
+
+    except KeyError as exc:
+        raise ValueError('missing required keyword argument: %s' % exc)
 
 
 class OutputEventFilter(object):
